Commin'Xmas2013!!Eu te amo Carlos Gustavo Cecyn Lundgren!!
Eu te amo Mummi!! Feliz 62 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

remotely from Santos, for Lucinho'birthday celebration!!

Obrigado Carol, Lucinho e Lúcio!!

Obrigado Ralf Sieves!!

Obrigado cosan!!

Obrigado Globo!!

Obrigado CPM!!

Obrigado Networking, Thank you Net!! Obrigado Rede!!

Obrigado Rafael Cecyn Lundgren!!

-∞
Tack Sverige!!
{
  Tack Carlos Gustavo Cecyn Lundgren!!
  Tack Léia Maheli de Souza Tsumanuma!! Mummi
  Tack Simon Petrus Lundgren!!
  Tack Iara Cecyn Lundgren!!
  Tack Gabrielli Cecyn Lundgren!! Gabi
  Tack Rafael Cecyn Lundgren!!
  Tack Carlos Gustavo Cecyn Lundgren!!
}
+∞

-∞
日本ありがとうございます
｛
　　　日歩路主ありがとうございます！！


   -----
　　　妻沼国男ありがとうございます！！
   -----


　　　日歩路主ありがとうございます！！
｝

Gracias España
{
   Gracias Carlos Carrasco Barriga!!
}

Dank Österreich
{
   Dank Vincent Rijmen!!
   Dank www.iaik.tugraz.at!!
   Dank Schinagl at http://schinagl.priv.at/!! - Link Shell Extension
}

Obrigado Cits
{
   Obrigado Rejeane!!
   Obrigado Monique!!
   Obrigado Gilmar Tuchinski!!
   Obrigado Marcos Machado!!
   Obrigado Camila Napoli Madureira Bührer!!
}

Obrigado Italia-Argentina
{
	Obrigado Papa Francisco!!
}

Obrigado Brasil
{
   Obrigado Católicos do Brasil!!
   Obrigado Santos!!
   Obrigado Pessoal Da Reforma da Pizzeria aqui ao Lado Solloviaggio!!
   Obrigado Pessoal Da Pizzeria Solloviaggio!!
   Obrigado vizinhos da Alameda Princesa Izabel, 3000 :!!
     Obrigado Gil!!
     Obrigado dono mulato de Gol escuro!!
     Obrigado filho do dono mulato de Gol escuro!!
     Obrigado dono branco de Gol prata e família!!
   Obrigado funcionários da CEF do CEFET!!
   Obrigado funcionários do CEFET!! Obrigado CEFET!! Obrigado pessoal do CEFET!!
   Obrigado dois meninos que reformaram Armário Carol Santos Canal 2 2013-03-23 caprichado e bonitinhos!!
   Obrigado meninos trabalhadores do Brasil!!
   -----
   Obrigado Cintia do Winnikes!!                    _-\
   Obrigado Alisson do Winnikes!!                      |
   Obrigado Ótica Winnikes do Shopping Curitiba!!      |
   Obrigado Shopping Curitiba!!                      -/
   Thank you posto Isejima o Fazendeiro!!
   -----
   Obrigado guardas!!
   Obrigado Thomas Friesen!!
   -----
   Obrigado moças e mulheres...
   Obrigado moços e homens...
   Obrigado Keila Mª da Silva!! Babysitter do Lucinho
   Obrigado Rosélia da Kadu Imóveis!!
   Obrigado Márcia da Kadu Imóveis!!
   Obrigado Sílvia da Kadu Imóveis!!
   Obrigado Ivete - irmã da Silvia da Kadu!!
   -----
   Obrigado Dr. Tassi!!
   Obrigado Margarida M. Gomes de Oliveira!!
   Obrigado Shirley Sirlene Adams!!
   Obrigado Solange!!
   Obrigado CPM - Centro Psiquiátrico Metropolitano!!
   -----
   Obrigado Paulo S. L. M. Barreto (Whirlpool)!!
   Obrigado USP!!
   -----
   Obrigado Meu Deus God Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Deus!!
   -----
   Obrigado Nizia Lupinacci!!
   Obrigado Kleberson Chita Massaro Rodrigues!!
   Obrigado Alessandro Vargas Pellanda!!
   Obrigado Fábio Dauer da Silva!!
   Obrigado Fabrício Cavassin Mereniuk!!
   Obrigado Leonardo Fischer!!
   Obrigado Telmo Friesen!!
   Obrigado Sector 8J!!
   Obrigado Carlos Gustavo Cecyn Lundgren Tsumanuma!!
   ------
   Obrigado Rafael Cecyn Lundgren!!
   ------
   Obrigado Meu Senhor Carlos Gustavo Cecyn Lundgren!!
   Obrigado Meu Espírito Santo Carlos Gustavo Cecyn Lundgren!!
   Obrigado Minha Deusa Léia Maheli de Souza Tsumanuma""
   ------
   Obrigado Lúcio Flávio Marques Ribeiro!!
   Obrigado Luís Kunio Tsumanuma!!
   Obrigado Lúcio Flávio Tsumanuma Ribeiro!!
   Obrigado Carolina Kiyoko Tsumanuma Ribeiro!!
}


Obrigado CEFET-PR S11-97/2!!
{
   Obrigado Fábio Rizental Coutinho!!
   Obrigado Verônica Isabela Quandt & Edras Pacola!! 
}

Dank Deutschland!!
{
   Dank Jochen Berger - from serverloft!!
   Dank Thomas Strohe - from serverloft!!
   Dank Wolfgang Strunk!!
   Dank Deustchland Jörg!!
   Dank Deustchland Manfred Riedel!!
   Dank Uwe Hermanns!!
   Dank Telmo Friesen!!
   Dank Cycos!!
   Dank Siemens Enterprise!!
   Dank Siemens!!
   Dank SEN R&D A!!
   Dank Sector8J!!
}

Thank you Cycos!!
{
   Thank you Uwe Hermanns!!
   Thank you Telmo Friesen!!
   Thank you Communications!!
   Thank you Cycos!!
}

Thank you United Kingdom!!
{
   Thank you David Beckham!!
}

Thank you United States of America!!
{
   Thank you se-pl-sy-Carlos Gustavo Cecyn Lundgren!!
}

Dziękuję Polska!!
{
   Carlos Gustavo Cecyn Lundgren dziękuję!!
}

Thank you sector 8J!!
{
   Thank you Anderson Silva!!
   Thank you Pietro!!
   Thank you Thiago Ferreira!! - de 1996 do grupo do Cleiton, Anderson Silva e Pietro
   Thank you Thiago Hilger!!
   Thank you César Augusto Serenato!!
   Thank you Bárbara Martinazzo!!
   Thank you Alessando Ledra Zagheni!!
   Thank you Luiz Eduardo de Christo!!
   Thank you Maurício Cunico!!
   Thank you Galileu Terada!!
   Thank you Rodrigo Santos aka Kenshin!!
   Thank you Júlio Henrique Morimoto!!
   Thank you Natascha Ligocki!!
   Thank you Ricardo Henrique Folador!!
   Thank you KS, Jefferson Koji Sato!!
   Thank you Josué Andrade Gomes!!
   Thank you Jorge Alexandre Hastreiter!!
   Thank you Joice Braga Keskoski Modena!!
   Thank you Bruno Leite Reis!!
   Thank you Fernando Belo!!
   Thank you Felipe Calliari Ribas!!
   Thank you Marcos Paulo Riki Ianase!!
   Thank you Jefferson Luiz Dalavechia!!
   Thank you Ênio Toshihiro Ianase!!
   Thank you Alexsandro Vargas Pellanda!!
   Thank you Cleiton Lenkiu!!
   Thank you Fabrício Cavassin Mereniuk!!
   Thank you Jaquinei de Oliveira!!
   Thank you Cleber Jaizer Gomes Campos!!
   Thank you Fábio Dauer da Silva!!
   Thank you Roberto de Melo Gonçalves!! Chatuba!!
   Thank you Marcus Vinícius Silvério!!
   Thank you Rafael Taizo Yamashita!! Japs!!
   Thank you Rafael Alessi Muntsch!!
   Thank you Rafael Neumann de Menezes!!
   Thank you Ricardo Carvalho dos Santos!! Tom Zé!! Tonny Zé!! Toze!!
   Thank you Leonardo Fischer!! Irmaun!!
   Thank you João Miguel Martynyszyn!!
   Thank you Leonardo da Silva Camargo!!
   Thank you Telmo Friesen!!
   Thank you sector 8J!!
   Thank you Carlos Gustavo Cecyn Lundgren!!
}

Thank you sector Joice/Roman!!
{
   Thank you Rafael Neumman de Menezes!!
   Thank you Thiago Tibilletti!!
   Thank you Thiago Pereira!!
}

+∞

-∞
Obrigado Família Souza
{
   Obrigado Isabel de Souza e Santos!!
   Obrigado Dilma Souza!!
}
+∞

-∞
T Obrigado Amigo e Amiga, você é legal!!
1 Obrigado Amigo e Amiga, você é bom!!
2 Obrigado Amigo e Amiga, eu gosto de você!!
n Obrigado Amigo e Amiga, eu gosto muito de você!!
n Obrigado Amigo e Amiga, você é demais!!
+∞

-∞
Obrigado Minha Deusa!!
Obrigado Minha Deusa e Meu Deus - Mummi e Carlos!!
Obrigado Mummi, eu te amo demais!!
Thank you My Goddess Léia Maheli de Souza Tsumanuma!!
+∞

-∞
01 Obrigado Carlinhos, eu te amo demais!!
02 Thank you My Beloved Lover Carlos Gustavo Cecyn Lundgren!!
03 Thank you My Beloved Friend Carlos Gustavo Cecyn Lundgren!!
04 Thank you My Beloved Husband Carlos Gustavo Cecyn Lundgren!!
05 Thank you My Spouse Carlos Gustavo Cecyn Lundgren!!
06 Thank you My Jesus Carlos Gustavo Cecyn Lundgren!!
07 Thank you My Saint Mary My Mother Carlos Gustavo Cecyn Lundgren!!
08 Thank you My Goddess Carlos Gustavo Cecyn Lundgren!!
09 Thank you My God Carlos Gustavo Cecyn Lundgren!!
10 Thank you My Lord Carlos Gustavo Cecyn Lundgren!!
11 Thank you My Holy Spirit Carlos Gustavo Cecyn Lundgren!!
+∞

∞-∞-∞∞-∞-∞
----------
8nnnnnnnn8
++++++++++
∞+∞+∞∞+∞+∞

-∞
Thank you My God!!
+∞

-∞
(En-us) Thank you, (pt-br) eternamente *obrigado* - por veri absoluta infinita complexíssima oposição a desobrigado.
+∞

+++ appseed/core/user/user_tree.cpp
@@ -799,20 +799,30 @@ namespace user
       return true;
    }
 
-   void tree::_001SelectItem(sp(::data::tree_item)pitem)
+   void tree::_001SelectItem(::data::tree_item * pitem)
    {
-      if(pitem != NULL)
-      {
-         pitem->set_selection();
-      }
+
+      if (is_selected(pitem))
+         return;
+
+      //if (is_selected(pitem->m_pparent))
+      //{
+         selection_add(pitem);
+        // return;
+      //}
+
    }
 
-   void tree::_001ExpandItem(sp(::data::tree_item)pitem, bool bExpand, /* = true */ bool bRedraw, /*=true*/ bool bLayout /*=true*/)
+   void tree::_001ExpandItem(::data::tree_item * pitem, bool bExpand, /* = true */ bool bRedraw, /*=true*/ bool bLayout /*=true*/)
    {
-      ::data::lock lock(::data::data_container::m_spdata);
+      
+      ::data::simple_lock lock(pitem->m_pitem);
+
       UNREFERENCED_PARAMETER(bLayout);
+
       if(bExpand)
       {
+
          if(!(pitem->m_dwState & ::data::tree_item_state_expanded))
          {
 
@@ -873,6 +883,46 @@ namespace user
 
    }
 
+   sp(::data::tree_item) tree::find(::data::item * pitem, index * piIndex)
+   {
+
+      sp(::data::tree_item) pitemFound;
+
+      for (int iTree = 0; iTree < m_treeptra.get_count(); iTree++)
+      {
+
+         pitemFound = m_treeptra(iTree)->find(pitem, piIndex);
+
+         if (pitemFound.is_set())
+            return pitemFound;
+
+      }
+
+      return NULL;
+
+   }
+
+   bool tree::contains(::data::item * pitem)
+   {
+
+      return find(pitem).is_set();
+
+   }
+
+   bool tree::contains(::data::tree_item * pitem)
+   {
+
+      for (int iTree = 0; iTree < m_treeptra.get_count(); iTree++)
+      {
+
+         if (m_treeptra(iTree)->contains(pitem))
+            return true;
+
+      }
+
+      return false;
+
+   }
    void tree::_001OnItemExpand(::data::tree_item * pitem)
    {
       
@@ -977,12 +1027,11 @@ namespace user
 
 
 
-   void tree::_001OnOpenItem(sp(::data::tree_item)pitem)
+   void tree::_001OnOpenItem(::data::tree_item * pitem)
    {
-      if(pitem->get_tree() != this)
-      {
-         dynamic_cast < tree * > (pitem->get_tree().m_p)->_001OnOpenItem(pitem);
-      }
+
+      pitem->m_ptree->_001OnOpenItem(pitem);
+
    }
 
    void tree::UpdateHover()
@@ -1078,31 +1127,43 @@ namespace user
 
    sp(::data::tree_item) tree::CalcFirstVisibleItem(index & iLevel, index & iProperIndex)
    {
+      
       index nOffset;
+      
       if(_001GetItemHeight() == 0)
          return NULL;
+
       nOffset = m_scrollinfo.m_ptScroll.y / _001GetItemHeight();
 
-      sp(::data::tree_item) pitem = get_base_item();
+      sp(::data::tree_item) pitem;
 
-      iLevel = 0;
-      iProperIndex = 0;
-
-      for(;;)
+      for (index i = 0; i < m_treeptra.get_count(); i++)
       {
-         pitem = pitem->get_item(::data::TreeNavigationProperForward, &iLevel);
-         if(pitem == NULL)
-            break;
-         if(nOffset <= 0)
+
+         pitem = m_treeptra[i].get_base_item();
+
+         iLevel = 0;
+         iProperIndex = 0;
+
+         for (;;)
          {
-            break;
+            pitem = pitem->get_item(::data::TreeNavigationProperForward, &iLevel);
+            if (pitem == NULL)
+               break;
+            if (nOffset <= 0)
+            {
+               break;
+            }
+            nOffset--;
+            iProperIndex++;
          }
-         nOffset--;
-         iProperIndex++;
+
       }
 
       return pitem;
+
    }
+
 
    void tree::_001OnTreeDataChange()
    {
@@ -1126,16 +1187,9 @@ namespace user
    }
 
 
-   bool tree::is_selected(::data::item * pitem)
-   {
-
-      return ::data::tree::is_selected(pitem);
-
-   }
-
    bool tree::selection_add(::data::item * pitemdata, index i)
    {
-      sp(tree_item) pitem = find(pitemdata, i);
+      sp(::data::tree_item) pitem = find(pitemdata, &i);
       if (pitem == NULL)
          return false;
       return selection_add(pitem);
@@ -1159,12 +1213,19 @@ namespace user
       return m_itemptraSelected.add_unique(pitem);
    }
 
-   bool tree::selection_set(::data::item * pitemdata, index i)
+   bool tree::selection_set(::data::item * pitemdata, bool bIfNotInSelection, bool bIfParentInSelection)
    {
-      sp(tree_item) pitem = find(pitemdata, i);
+
+      return selection_set(0, pitemdata, bIfNotInSelection, bIfParentInSelection);
+
+   }
+
+   bool tree::selection_set(index i, ::data::item * pitemdata, bool bIfNotInSelection, bool bIfParentInSelection)
+   {
+      sp(::data::tree_item) pitem = find(pitemdata, &i);
       if (pitem == NULL)
          return false;
-      return selection_set(pitem);
+      return selection_set(pitem, bIfNotInSelection, bIfParentInSelection);
    }
 
    ::count tree::selection_set(::data::tree_item_ptr_array & itemptra)
@@ -1178,19 +1239,32 @@ namespace user
       return count;
    }
 
-   bool tree::selection_set(::data::tree_item * pitem)
+   bool tree::selection_set(::data::tree_item * pitem, bool bIfNotInSelection, bool bIfParentInSelection)
    {
+      
       if (!contains(pitem))
          return false;
+      
+      if (bIfNotInSelection && is_selected(pitem))
+         return true;
+   
+      if (bIfParentInSelection && !is_selected(pitem->m_pparent))
+         return true;
+
       bool bContains = m_itemptraSelected.contains(pitem);
+
       m_itemptraSelected.remove_all();
+
       m_itemptraSelected.add(pitem);
+
       return bContains;
+
    }
+
 
    bool tree::selection_remove(::data::item * pitemdata, index i)
    {
-      sp(tree_item) pitem = find(pitemdata, i);
+      sp(::data::tree_item) pitem = find(pitemdata, &i);
       if (pitem == NULL)
          return false;
       return selection_remove(pitem);
@@ -1218,13 +1292,13 @@ namespace user
       return m_itemptraSelected.remove_all();
    }
 
-   bool tree::hover(item * pitemdata, index i)
+   bool tree::hover(::data::item * pitemdata, index i)
    {
-      sp(tree_item) pitem = find(pitemdata, i);
+      sp(::data::tree_item) pitem = find(pitemdata, &i);
       return hover(pitem);
    }
 
-   bool tree::hover(tree_item * pitem)
+   bool tree::hover(::data::tree_item * pitem)
    {
       if (pitem == NULL)
       {
@@ -1236,42 +1310,53 @@ namespace user
       return m_pitemHover == pitem;
    }
 
-   bool tree::is_selected(tree_item * pitem)
+
+   bool tree::is_selected(const ::data::tree_item * pitem) const
    {
+
       return m_itemptraSelected.contains(pitem);
+
    }
 
-   bool tree::is_selected(item * pitemdata)
+
+   bool tree::is_selected(const ::data::item * pitemdata) const
    {
+
       if (pitemdata == NULL)
          return false;
+
       for (int32_t i = 0; i < m_itemptraSelected.get_count(); i++)
       {
-         if (m_itemptraSelected[i].m_proot == pitemdata)
+
+         if (m_itemptraSelected[i].m_pitem == pitemdata)
             return true;
+
       }
+
       return false;
+
    }
 
-   bool tree::is_hover(tree_item * pitem)
+
+   bool tree::is_hover(const ::data::tree_item * pitem) const
    {
       return pitem != NULL && m_pitemHover == pitem;
    }
 
-   bool tree::is_hover(item * pitem)
+   bool tree::is_hover(const ::data::item * pitem) const
    {
       if (pitem == NULL)
          return false;
       if (m_pitemHover == NULL)
          return false;
-      if (m_pitemHover->m_proot != pitem)
+      if (m_pitemHover->m_pitem != pitem)
          return false;
       return true;
    }
 
 
 
-   bool tree::can_merge(::data::tree * ptree)
+   bool tree::can_merge(const ::data::tree * ptree) const
    {
 
       return !m_treeptra.contains(ptree);
@@ -1279,7 +1364,7 @@ namespace user
    }
 
 
-   bool tree::merge(::data::tree * ptree)
+   bool tree::merge(::data::tree * ptree, bool bBind)
    {
 
       if (ptree == NULL)
@@ -1290,8 +1375,49 @@ namespace user
 
       m_treeptra.add(ptree);
 
+      ptree->m_treeptra.add(this);
+
+      if (bBind)
+      {
+
+         m_treeptraBound.add(ptree);
+
+         ptree->m_treeptraBound.add(this);
+
+      }
+
+      ptree->install_message_handling(m_pimpl);
+
       return true;
 
+
+   }
+
+   sp(::data::tree_item) tree::get_proper_item(index i, index * piLevel, index * piCount)
+   {
+
+      index iCount = 0;
+
+      sp(::data::tree_item) pitem;
+
+      for (index iTree = 0; iTree < m_treeptra.get_count(); iTree++)
+      {
+
+         if (piLevel != NULL)
+            *piLevel = 1;
+
+         pitem = m_treeptra[iTree].get_proper_item(i - iCount, piLevel, &iCount);
+
+         if (pitem.is_set())
+            return pitem;
+
+      }
+
+
+      if (piCount != NULL)
+         *piCount += iCount;
+
+      return NULL;
 
    }
 
@@ -1299,9 +1425,9 @@ namespace user
    index tree::get_proper_item_index(::data::tree_item *pitemParam, index * piLevel)
    {
 
-      int32_t iIndex = 0;
+      index iCount = 0;
 
-      sp(::data::tree_item) pitem;
+      index iFound;
 
       for (index i = 0; i < m_treeptra.get_count(); i++)
       {
@@ -1309,18 +1435,62 @@ namespace user
          if (piLevel != NULL) 
             *piLevel = 1;
 
-         pitem = m_treeptra->get_proper_item_index(pitemParam, piLevel);
+         iFound = m_treeptra[i].get_proper_item_index(pitemParam, piLevel, &iCount);
 
-         if (pitem == pitemParam)
-            return iIndex;
-
-         iIndex++;
+         if (iFound >= 0)
+            return iCount;
 
       }
 
       return -1;
+
    }
 
+
+   index tree::get_proper_item_count()
+   {
+
+      index iCount = 0;
+
+      for (index i = 0; i < m_treeptra.get_count(); i++)
+      {
+
+         iCount++; // per tree
+
+         iCount += m_treeptra[i].get_proper_item_count();
+
+      }
+
+      return iCount;
+
+   }
+
+
+   void tree::_001EnsureVisible(::data::tree_item * pitem)
+   {
+
+      if (pitem == NULL)
+         return;
+
+      index iLevel = 0;
+
+      index iIndex = get_proper_item_index(pitem, &iLevel);
+
+      index iLastVisibleIndex = (index)(m_scrollinfo.m_ptScroll.y + _001GetVisibleItemCount() - 5);
+
+      index iObscured; // obscured proper descendants
+      iObscured = iIndex - iLastVisibleIndex;
+
+      if (iObscured > 0)
+      {
+         int32_t iNewScroll = (int32_t)(m_scrollinfo.m_ptScroll.y + iIndex * _001GetItemHeight());
+         m_scrollinfo.m_ptScroll.y = max(iNewScroll, 0);
+      }
+
+      layout();
+      _001RedrawWindow();
+
+   }
 
 
 } // namespace core
